+++ src/test/formatter_component_tags.test.ts
@@ -528,4 +528,398 @@ suite('Component Tags', () => {
 `;
         assert.strictEqual(formatBladeString(input), output);
     });
+
+    test('custom html tags that have the same name as void tags can be formatted', () => {
+        const input = `
+<object data="audi.wav">
+  <param name="autoplay" value="true">
+</object>
+<Param something="here">
+    <Param />
+    <param name="autoplay" value="true">
+</Param>
+<audio controls>
+  <source src="1.ogg" type="audio/ogg">
+  <source src="2.mp3" type="audio/mpeg">
+  Your browser does not support the audio element.
+  
+  <track src="1.vtt" kind="subtitles" srclang="en" label="1">
+  <track src="2.vtt" kind="subtitles" srclang="no" label="2">
+</audio>
+<p>Test<wbr>content</p>
+<Wbr />
+<Wbr>
+
+</Wbr>
+<Track>
+    The track.
+    <Track />
+</Track>
+
+<Source src="something.wav" />
+<Source>
+<object data="audi.wav">
+  <param name="autoplay" value="true">
+</object>
+<Param something="here">
+    <Param />
+    <param name="autoplay" value="true">
+</Param>
+<audio controls>
+  <source src="1.ogg" type="audio/ogg">
+  <source src="2.mp3" type="audio/mpeg">
+  Your browser does not support the audio element.
+</audio>
+</Source>
+<hr>
+<hr />
+<Hr repeat="2" />
+<Hr>
+    Hello!
+</Hr>
+<img src="" />
+<img src="">
+<Img type="something">
+Hello, there!
+</Img>
+<input type="text">
+<input type="text"/>
+<Input type="text" />
+<Input type="text">
+    Placeholder text.
+</Input>
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<Meta>
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <Description>Hello!</Description>
+</Meta>
+<command type="checklist" label="Check List">Check List</command>
+<Command something="here">
+<map name="workmap">
+  <area shape="rect" coords="34,44,270,350" alt="Computer" href="computer.htm">
+  <area shape="rect" coords="290,172,333,250" alt="Phone" href="phone.htm">
+  <area shape="circle" coords="337,300,44" alt="Cup of coffee" href="coffee.htm">
+</map>
+<embed type="image/jpg" src="pic.jpg" width="300" height="200">
+<Area width="100" height="100" />
+<Embed src="something.svg">
+
+<command type="checklist" label="Check List">Check List</command>
+<Command something="here">
+<map name="workmap">
+  <area shape="rect" coords="34,44,270,350" alt="Computer" href="computer.htm">
+  <area shape="rect" coords="290,172,333,250" alt="Phone" href="phone.htm">
+  <area shape="circle" coords="337,300,44" alt="Cup of coffee" href="coffee.htm">
+</map>
+<embed type="image/jpg" src="pic.jpg" width="300" height="200">
+<Area width="100" height="100" />
+<base href="https://www.w3schools.com/" target="_blank">
+<Area width="100" height="100"/>
+<Area width="100" height="100">
+    Something.
+</Area>
+</Command>
+</Embed>
+
+<map name="workmap">
+  <area shape="rect" coords="34,44,270,350" alt="Computer" href="computer.htm">
+  <area shape="rect" coords="290,172,333,250" alt="Phone" href="phone.htm">
+  <area shape="circle" coords="337,300,44" alt="Cup of coffee" href="coffee.htm">
+</map>
+
+<Area width="100" height="100" />
+
+<base href="https://www.w3schools.com/" target="_blank">
+<Area width="100" height="100"/>
+<Area width="100" height="100">
+    Something.
+</Area>
+
+<br>
+<br />
+<table>
+  <colgroup>
+    <col span="2" style="background-color:red">
+    <col style="background-color:yellow">
+  </colgroup>
+</table>
+
+<Col span="12">
+    <Row>
+
+    </Row>
+</Col>
+
+<Br repeat="2" />
+
+<Br repeat="3">
+<base href="https://www.w3schools.com/" target="_blank">
+<Area width="100" height="100"/>
+<Area width="100" height="100">
+    Something.
+</Area>
+</Br>
+
+<Base width="100" height="100">
+    Something.
+</Base>
+
+@if ($something)
+<link rel="">
+<Link href="{{ route('profile.show') }}" class="text-sm text-gray-600 underline hover:text-gray-900">
+    {{ __('Edit Profile') }}
+    <Base width="100" height="100">
+    Something.
+    <map name="workmap">
+  <area shape="rect" coords="34,44,270,350" alt="Computer" href="computer.htm">
+  <area shape="rect" coords="290,172,333,250" alt="Phone" href="phone.htm">
+  <area shape="circle" coords="337,300,44" alt="Cup of coffee" href="coffee.htm">
+</map>
+
+<Area width="100" height="100" />
+
+<base href="https://www.w3schools.com/" target="_blank">
+<Area width="100" height="100">
+    Something.
+</Area>
+</Base>
+</Link>
+
+@endif
+
+<base href="https://www.w3schools.com/" target="_blank">
+<link rel="">
+<Link href="{{ route('profile.show') }}" class="text-sm text-gray-600 underline hover:text-gray-900">
+    {{ __('Edit Profile') }}
+</Link>
+
+
+`;
+        const expected = `<object data="audi.wav">
+    <param name="autoplay" value="true" />
+</object>
+<Param something="here">
+    <Param />
+    <param name="autoplay" value="true" />
+</Param>
+<audio controls>
+    <source src="1.ogg" type="audio/ogg" />
+    <source src="2.mp3" type="audio/mpeg" />
+    Your browser does not support the audio element.
+
+    <track src="1.vtt" kind="subtitles" srclang="en" label="1" />
+    <track src="2.vtt" kind="subtitles" srclang="no" label="2" />
+</audio>
+<p>
+    Test
+    <wbr />
+    content
+</p>
+<Wbr />
+<Wbr></Wbr>
+<Track>
+    The track.
+    <Track />
+</Track>
+
+<Source src="something.wav" />
+<Source>
+    <object data="audi.wav">
+        <param name="autoplay" value="true" />
+    </object>
+    <Param something="here">
+        <Param />
+        <param name="autoplay" value="true" />
+    </Param>
+    <audio controls>
+        <source src="1.ogg" type="audio/ogg" />
+        <source src="2.mp3" type="audio/mpeg" />
+        Your browser does not support the audio element.
+    </audio>
+</Source>
+<hr />
+<hr />
+<Hr repeat="2" />
+<Hr>Hello!</Hr>
+<img src="" />
+<img src="" />
+<Img type="something">
+    Hello, there!
+</Img>
+<input type="text" />
+<input type="text" />
+<Input type="text" />
+<Input type="text">
+    Placeholder text.
+</Input>
+<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+<Meta>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <Description>Hello!</Description>
+</Meta>
+<command type="checklist" label="Check List">Check List</command>
+<Command something="here">
+    <map name="workmap">
+        <area
+            shape="rect"
+            coords="34,44,270,350"
+            alt="Computer"
+            href="computer.htm"
+        />
+        <area
+            shape="rect"
+            coords="290,172,333,250"
+            alt="Phone"
+            href="phone.htm"
+        />
+        <area
+            shape="circle"
+            coords="337,300,44"
+            alt="Cup of coffee"
+            href="coffee.htm"
+        />
+    </map>
+    <embed type="image/jpg" src="pic.jpg" width="300" height="200" />
+    <Area width="100" height="100" />
+    <Embed src="something.svg">
+        <command type="checklist" label="Check List">Check List</command>
+        <Command something="here">
+            <map name="workmap">
+                <area
+                    shape="rect"
+                    coords="34,44,270,350"
+                    alt="Computer"
+                    href="computer.htm"
+                />
+                <area
+                    shape="rect"
+                    coords="290,172,333,250"
+                    alt="Phone"
+                    href="phone.htm"
+                />
+                <area
+                    shape="circle"
+                    coords="337,300,44"
+                    alt="Cup of coffee"
+                    href="coffee.htm"
+                />
+            </map>
+            <embed type="image/jpg" src="pic.jpg" width="300" height="200" />
+            <Area width="100" height="100" />
+            <base href="https://www.w3schools.com/" target="_blank" />
+            <Area width="100" height="100" />
+            <Area width="100" height="100">
+                Something.
+            </Area>
+        </Command>
+    </Embed>
+
+    <map name="workmap">
+        <area
+            shape="rect"
+            coords="34,44,270,350"
+            alt="Computer"
+            href="computer.htm"
+        />
+        <area
+            shape="rect"
+            coords="290,172,333,250"
+            alt="Phone"
+            href="phone.htm"
+        />
+        <area
+            shape="circle"
+            coords="337,300,44"
+            alt="Cup of coffee"
+            href="coffee.htm"
+        />
+    </map>
+
+    <Area width="100" height="100" />
+
+    <base href="https://www.w3schools.com/" target="_blank" />
+    <Area width="100" height="100" />
+    <Area width="100" height="100">
+        Something.
+    </Area>
+
+    <br />
+    <br />
+    <table>
+        <colgroup>
+            <col span="2" style="background-color: red" />
+            <col style="background-color: yellow" />
+        </colgroup>
+    </table>
+
+    <Col span="12">
+        <Row></Row>
+    </Col>
+
+    <Br repeat="2" />
+
+    <Br repeat="3">
+        <base href="https://www.w3schools.com/" target="_blank" />
+        <Area width="100" height="100" />
+        <Area width="100" height="100">
+            Something.
+        </Area>
+    </Br>
+
+    <Base width="100" height="100">
+        Something.
+    </Base>
+
+    @if ($something)
+        <link rel="" />
+        <Link
+            href="{{ route("profile.show") }}"
+            class="text-sm text-gray-600 underline hover:text-gray-900"
+        >
+            {{ __("Edit Profile") }}
+            <Base width="100" height="100">
+                Something.
+                <map name="workmap">
+                    <area
+                        shape="rect"
+                        coords="34,44,270,350"
+                        alt="Computer"
+                        href="computer.htm"
+                    />
+                    <area
+                        shape="rect"
+                        coords="290,172,333,250"
+                        alt="Phone"
+                        href="phone.htm"
+                    />
+                    <area
+                        shape="circle"
+                        coords="337,300,44"
+                        alt="Cup of coffee"
+                        href="coffee.htm"
+                    />
+                </map>
+
+                <Area width="100" height="100" />
+
+                <base href="https://www.w3schools.com/" target="_blank" />
+                <Area width="100" height="100">
+                    Something.
+                </Area>
+            </Base>
+        </Link>
+    @endif
+
+    <base href="https://www.w3schools.com/" target="_blank" />
+    <link rel="" />
+    <Link
+        href="{{ route("profile.show") }}"
+        class="text-sm text-gray-600 underline hover:text-gray-900"
+    >
+        {{ __("Edit Profile") }}
+    </Link>
+</Command>
+`;
+        assert.strictEqual(formatBladeString(input), expected);
+    });
 });